# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/files/raw/{key}": {
      "get": {
        "operationId": "FilesController_raw",
        "parameters": [
          {
            "name": "key",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Serve a locally-stored file (dev storage)",
        "tags": [
          "files"
        ]
      }
    },
    "/api/files/upload": {
      "post": {
        "operationId": "FilesController_upload",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  },
                  "tags": {
                    "type": "string",
                    "description": "Comma-separated memory tags",
                    "example": "brand_guideline,compliance"
                  },
                  "importance": {
                    "type": "string",
                    "description": "Memory importance tier 0..1",
                    "example": "0.75"
                  },
                  "locked": {
                    "type": "string",
                    "description": "Lock from auto-importance updates (true/false)",
                    "example": "false"
                  }
                },
                "required": [
                  "file"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "File stored and ingested into memory"
          }
        },
        "summary": "Upload a document into RAG memory (PDF/DOCX/TXT/MD)",
        "tags": [
          "files"
        ]
      }
    },
    "/api/files/{id}": {
      "patch": {
        "operationId": "FilesController_updateMeta",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Update a document’s importance / lock / markdown text (re-indexes memory)",
        "tags": [
          "files"
        ]
      },
      "get": {
        "operationId": "FilesController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get a single uploaded document",
        "tags": [
          "files"
        ]
      }
    },
...
    "/api/campaigns": {
      "post": {
        "operationId": "CampaignsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCampaignDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Campaign created"
          }
        },
        "summary": "Create a campaign",
        "tags": [
          "campaigns"
        ]
      },
      "get": {
        "operationId": "CampaignsController_findAll",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "draft",
                "active",
                "paused",
                "completed"
              ],
              "type": "string"
            }
          },
          {
            "name": "health",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "healthy",
                "warning",
                "at_risk",
                "critical"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List campaigns (optionally filter by status/health)",
        "tags": [
          "campaigns"
        ]
      }
    },
    "/api/campaigns/{id}": {
      "get": {
        "operationId": "CampaignsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get a campaign with posts, ads, and open recommendations",
        "tags": [
          "campaigns"
        ]
      },
      "patch": {
        "operationId": "CampaignsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCampaignDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Update a campaign",
        "tags": [
          "campaigns"
        ]
      }
    },
    "/api/campaigns/{id}/summary": {
      "get": {
        "operationId": "CampaignsController_summary",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Campaign summary: goal progress + AI narrative",
        "tags": [
          "campaigns"
        ]
      }
    },
    "/api/campaigns/{id}/progress": {
      "get": {
        "operationId": "CampaignsController_progress",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Goal attainment and health snapshot",
        "tags": [
          "campaigns"
        ]
      }
    },
    "/api/posts": {
      "post": {
        "operationId": "PostsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePostDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Post created"
          }
        },
        "summary": "Create a post (content calendar entry)",
        "tags": [
          "posts"
        ]
      },
      "get": {
        "operationId": "PostsController_findAll",
        "parameters": [
          {
            "name": "campaignId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "draft",
                "review",
                "approved",
                "scheduled",
                "published",
                "analyzing",
                "underperforming",
                "completed"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List posts (optionally by campaign/status) — content calendar feed",
        "tags": [
          "posts"
        ]
      }
    },
...
    "/api/posts/{postId}/comments": {
      "get": {
        "operationId": "CommentsController_list",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List comments + agent suggestions on a post",
        "tags": [
          "comments"
        ]
      },
      "post": {
        "operationId": "CommentsController_create",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Add a human comment or suggestion to a post",
        "tags": [
          "comments"
        ]
      }
    },
    "/api/posts/{postId}/comments/{commentId}/accept": {
      "post": {
        "operationId": "CommentsController_accept",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "commentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Accept a suggestion (applies the edit + records learning)",
        "tags": [
          "comments"
        ]
      }
    },
    "/api/posts/{postId}/comments/{commentId}/choose": {
      "post": {
        "operationId": "CommentsController_choose",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "commentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Pick one of a suggestion’s options — the agent applies it",
        "tags": [
          "comments"
        ]
      }
    },
    "/api/posts/{postId}/comments/{commentId}/reject": {
      "post": {
        "operationId": "CommentsController_reject",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "commentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Reject a suggestion (records learning)",
        "tags": [
          "comments"
        ]
      }
    },
    "/api/posts/{postId}/comments/{commentId}/resolve": {
      "post": {
        "operationId": "CommentsController_resolve",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "commentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Resolve a comment thread",
        "tags": [
          "comments"
        ]
      }
    },
    "/api/posts/{id}/copy-edit": {
      "post": {
        "operationId": "PostCopyEditController_copyEdit",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Human copy edit — edits a draft in place, or (if the post is live) saves it to the memory bank and spins a new draft variant instead of mutating it",
        "tags": [
          "posts"
        ]
      }
    },
...
  },
  "info": {
    "title": "Hiive Agentic Marketing OS",
    "description": "Mission-driven, multi-agent marketing operating system: RAG memory, campaign intelligence, swarm simulation, autonomous monitoring, human-in-the-loop approvals, and a continuous learning flywheel.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateCampaignDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Sell-side Founder Liquidity Campaign"
          },
          "objective": {
            "type": "string",
            "example": "Increase inbound sellers of pre-IPO shares"
          },
          "audience": {
            "type": "string",
            "example": "Startup employees, founders, early investors"
          },
          "channels": {
            "example": [
              "LinkedIn",
              "Email",
              "X"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "budget": {
            "type": "number",
            "example": 25000
          },
          "goals": {
            "type": "object",
            "example": {
              "impressions": 50000,
              "clicks": 1500,
              "leads": 100
            },
            "description": "Goal targets (leads maps to conversions)"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "active",
              "paused",
              "completed"
            ]
          },
          "startDate": {
            "type": "string",
            "example": "2026-06-01T00:00:00.000Z"
          },
          "endDate": {
            "type": "string"
          },
          "missionId": {
            "type": "string",
            "description": "Owning mission id"
          }
        },
        "required": [
          "name",
          "objective",
          "audience",
          "goals"
        ]
      },
      "UpdateCampaignDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Sell-side Founder Liquidity Campaign"
          },
          "objective": {
            "type": "string",
            "example": "Increase inbound sellers of pre-IPO shares"
          },
          "audience": {
            "type": "string",
            "example": "Startup employees, founders, early investors"
          },
          "channels": {
            "example": [
              "LinkedIn",
              "Email",
              "X"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "budget": {
            "type": "number",
            "example": 25000
          },
          "goals": {
            "type": "object",
            "example": {
              "impressions": 50000,
              "clicks": 1500,
              "leads": 100
            },
            "description": "Goal targets (leads maps to conversions)"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "active",
              "paused",
              "completed"
            ]
          },
          "startDate": {
            "type": "string",
            "example": "2026-06-01T00:00:00.000Z"
          },
          "endDate": {
            "type": "string"
          },
          "missionId": {
            "type": "string",
            "description": "Owning mission id"
          }
        }
      },
      "CreatePostDto": {
        "type": "object",
        "properties": {
          "campaignId": {
            "type": "string",
            "example": "campaign id"
          },
          "platform": {
            "type": "string",
            "example": "LinkedIn"
          },
          "copy": {
            "type": "string",
            "example": "Thinking about liquidity before your IPO? ..."
          },
          "mediaUrl": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "review",
              "approved",
              "scheduled",
              "published",
              "analyzing",
              "underperforming",
              "completed"
            ]
          },
          "scheduledAt": {
            "type": "string",
            "example": "2026-06-10T09:00:00.000Z"
          }
        },
        "required": [
          "campaignId",
          "platform",
          "copy"
        ]
      },
...
      "CreateCommentDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "comment",
              "suggestion"
            ]
          },
          "body": {
            "type": "string",
            "example": "This CTA is too vague."
          },
          "quotedText": {
            "type": "string",
            "description": "Highlighted span the comment refers to"
          },
          "rangeStart": {
            "type": "number"
          },
          "rangeEnd": {
            "type": "number"
          },
          "suggestedText": {
            "type": "string",
            "description": "Replacement text for a suggestion"
          }
        },
        "required": [
          "type",
          "body"
        ]
      },
...
    }
  }
}